# openspec/changes/phase-4-backend-and-api-migration/contracts/JB2026.Api.swagger.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JB2026.Api",
    "description": "Phase 4 migrated API surface",
    "version": "v1"
  },
  "paths": {
...
    "/api/v2/jobs/range": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "startOn",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/jobs/{id}/details": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/quotations": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "startOn",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/quotations/search/{keyword}": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationListItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/quotations/{id}/pdf": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JobAttachmentResponse": {
        "required": [
          "contentType",
          "fileName",
          "length"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "JobDetailResponse": {
        "required": [
          "attachments",
          "customerName",
          "customerRef",
          "orderedBy",
          "orderedOn",
          "orderId",
          "orderNumber",
          "orderTitle",
          "paymentTerms",
          "qty",
          "remarks",
          "requiredOn",
          "status",
          "styleTitles"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerRef": {
            "type": "string",
            "nullable": true
          },
          "orderTitle": {
            "type": "string",
            "nullable": true
          },
          "orderedBy": {
            "type": "string",
            "nullable": true
          },
          "orderedOn": {
            "type": "string",
            "format": "date-time"
          },
          "requiredOn": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "qty": {
            "type": "number",
            "format": "double"
          },
          "paymentTerms": {
            "type": "string",
            "nullable": true
          },
          "remarks": {
            "type": "string",
            "nullable": true
          },
          "styleTitles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobAttachmentResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "JobListItemResponse": {
        "required": [
          "customerName",
          "customerRef",
          "orderedBy",
          "orderedOn",
          "orderId",
          "orderNumber",
          "orderTitle",
          "qty",
          "requiredOn",
          "status"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerRef": {
            "type": "string",
            "nullable": true
          },
          "orderTitle": {
            "type": "string",
            "nullable": true
          },
          "orderedBy": {
            "type": "string",
            "nullable": true
          },
          "orderedOn": {
            "type": "string",
            "format": "date-time"
          },
          "requiredOn": {
            "type": "string",
            "format": "date-time"
          },
          "qty": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "QuotationListItemResponse": {
        "required": [
          "customerName",
          "headerId",
          "machineType",
          "materialCost",
          "materialName",
          "printsColor",
          "printsQty",
          "printsSize",
          "printTitle",
          "quotedBy",
          "quotedOn",
          "quoteNumber",
          "quoteNumberIndex",
          "quoteNumberIndexPair",
          "status",
          "totalCostA",
          "unitCostA"
        ],
        "type": "object",
        "properties": {
          "headerId": {
            "type": "string",
            "format": "uuid"
          },
          "machineType": {
            "type": "string",
            "nullable": true
          },
          "quoteNumber": {
            "type": "integer",
            "format": "int32"
          },
          "quoteNumberIndex": {
            "type": "integer",
            "format": "int32"
          },
          "quoteNumberIndexPair": {
            "type": "string",
            "nullable": true
          },
          "quotedOn": {
            "type": "string",
            "format": "date-time"
          },
          "quotedBy": {
            "type": "string",
            "nullable": true
          },
          "approvedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "approvedBy": {
            "type": "string",
            "nullable": true
          },
          "printTitle": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "printsSize": {
            "type": "string",
            "nullable": true
          },
          "printsColor": {
            "type": "string",
            "nullable": true
          },
          "printsQty": {
            "type": "number",
            "format": "double"
          },
          "materialName": {
            "type": "string",
            "nullable": true
          },
          "materialCost": {
            "type": "number",
            "format": "double"
          },
          "totalCostA": {
            "type": "number",
            "format": "double"
          },
          "unitCostA": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Provide a valid JWT access token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}